Basic Python example
Expose PassManager, PassContext to Python and add a simple example running a pass.

Also rename the existing "passmanager" to "base_passmanager" where applicable
