+++ Cargo.toml
@@ -40,12 +40,15 @@ rand = "0.10"
 rand_pcg = "0.10"
 rand_distr = "0.6"
 num-traits = "0.2"
-uuid = { version = "1.23", features = ["v4", "fast-rng"], default-features = false }
+uuid = { version = "1.23", features = [
+    "v4",
+    "fast-rng",
+], default-features = false }
 anyhow = "1.0"
 binrw = "0.15"
 cbindgen = "0.29.4"
 mimalloc = "0.1.52"
-npyz = {version = "0.9.1", features = ["complex"]}
+npyz = { version = "0.9.1", features = ["complex"] }
 
 # Most of the crates don't need the feature `extension-module`, since only `qiskit-pyext` builds an
 # actual C extension (the feature disables linking in `libpython`, which is forbidden in Python
@@ -67,7 +70,8 @@ qiskit-qasm3 = { path = "crates/qasm3" }
 qiskit-qpy = { path = "crates/qpy" }
 qiskit-transpiler = { path = "crates/transpiler" }
 qiskit-quantum-info = { path = "crates/quantum_info" }
-qiskit-synthesis = {path = "crates/synthesis" }
+qiskit-synthesis = { path = "crates/synthesis" }
+qiskit-passmanager = { path = "crates/passmanager" }
 qiskit-util = { path = "crates/util" }
 
 [workspace.lints.clippy]
