+++ crates/pyext/Cargo.toml
@@ -20,7 +20,13 @@ workspace = true
 # `libpython` is *not* linked, and if the feature isn't present, then it is.  To test the Rust
 # crates as standalone binaries, executables, we need `libpython` to be linked in, so we make the
 # feature a default, and run `cargo test --no-default-features` to turn it off.
-cache_pygates = ["qiskit-circuit/cache_pygates", "qiskit-accelerate/cache_pygates", "qiskit-transpiler/cache_pygates", "qiskit-cext/cache_pygates", "qiskit-qpy/cache_pygates"]
+cache_pygates = [
+    "qiskit-circuit/cache_pygates",
+    "qiskit-accelerate/cache_pygates",
+    "qiskit-transpiler/cache_pygates",
+    "qiskit-cext/cache_pygates",
+    "qiskit-qpy/cache_pygates",
+]
 
 [build-dependencies]
 anyhow.workspace = true
@@ -41,4 +47,5 @@ qiskit-cext-vtable = { workspace = true, features = ["python_binding", "addr"] }
 qiskit-transpiler.workspace = true
 qiskit-quantum-info = { workspace = true, features = ["python"] }
 qiskit-synthesis.workspace = true
+qiskit-passmanager.workspace = true
 mimalloc = { workspace = true, optional = true }
