+++ qiskit/passmanager/pass_interface.py
@@ -1,0 +1,22 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2026.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at https://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+from abc import ABC, abstractmethod
+from typing import Any
+from qiskit._accelerate.passmanager import PassContext
+
+
+class Pass(ABC):
+    """The pass interface."""
+
+    @abstractmethod
+    def run(self, ir: Any, context: PassContext) -> Any: ...
